+++ ydsz-pmis-backend/ydsz-pmis-project/src/main/java/com/njydsz/pmis/project/controller/ProjectChangeController.java
@@ -175,7 +175,7 @@ public class ProjectChangeController {
     @Operation(summary = "统计项目重大变更数")
     @PrePermission("project:change:list")
     @GetMapping("/major-count/{initiationId}")
-    public Result<Long> countMajor(@PathVariable String initiationId) {
+    public Result<Integer> countMajor(@PathVariable String initiationId) {
         return Result.ok(service.countMajorByInitiation(initiationId));
     }
